[git] Reordering functions and types in git pkg
